Erm inconsistency

+++ example/include/SSGE/SSGE_objtemplate.h
@@ -19,7 +19,7 @@ extern "C" {
  * \return The object template id
  * \note The object template is stored internally and can be accessed by its name or its id
  */
-SSGEDECL uint32_t SSGE_Template_Create(char *name, SSGE_Texture *texture, int width, int height, bool hitbox, void (*destroyData)(void *));
+SSGEDECL uint32_t SSGE_Template_Create(char *name, SSGE_Texture *texture, uint16_t width, uint16_t height, bool hitbox, void (*destroyData)(void *));
 
 /**
  * Gets an object template by id
